Проектная работа 11

+++ src/utils/Api.js
@@ -47,18 +47,18 @@ class Api {
     }).then(this._checkResponse);
   }
 
-  addLike(id) {
-    return fetch(this._baseUrl + `/cards/${id}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    }).then(this._checkResponse);
-  }
-
-  deleteLike(id) {
-    return fetch(this._baseUrl + `/cards/${id}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    }).then(this._checkResponse);
+  changeLikeStatus(id, isLiked) {
+    if (isLiked) {
+      return fetch(this._baseUrl + `/cards/${id}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      }).then(this._checkResponse);
+    } else {
+      return fetch(this._baseUrl + `/cards/${id}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      }).then(this._checkResponse);
+    }
   }
 
   removeCard(id) {
